[Filesystem] delete files on entities delete
