+++ src/main/core/Resources/config/services/crud.yml
@@ -11,6 +11,7 @@ services:
             - '@Claroline\AppBundle\Persistence\ObjectManager'
             - '@Claroline\AppBundle\API\Crud'
             - '@Claroline\CoreBundle\Manager\Workspace\WorkspaceManager'
+            - '@Claroline\CoreBundle\Manager\FileManager'
             - '@Claroline\CoreBundle\Manager\Tool\ToolManager'
             - '@Claroline\CoreBundle\Manager\ResourceManager'
             - '@Claroline\CoreBundle\Manager\Organization\OrganizationManager'
@@ -35,14 +36,16 @@ services:
             - '@Claroline\CoreBundle\Manager\Workspace\WorkspaceManager'
             - '@Icap\NotificationBundle\Manager\NotificationUserParametersManager'
             - '@Claroline\AppBundle\Event\StrictDispatcher'
+            - '@Claroline\CoreBundle\Manager\FileManager'
         tags:
             - { name: kernel.event_listener, event: crud_pre_create_object_claroline_corebundle_entity_user, method: preCreate }
             - { name: kernel.event_listener, event: crud_post_create_object_claroline_corebundle_entity_user, method: postCreate }
-            - { name: kernel.event_listener, event: crud_pre_delete_object_claroline_corebundle_entity_user, method: preDelete }
             - { name: kernel.event_listener, event: crud_pre_update_object_claroline_corebundle_entity_user, method: preUpdate }
             - { name: kernel.event_listener, event: crud_post_update_object_claroline_corebundle_entity_user, method: postUpdate }
             - { name: kernel.event_listener, event: crud_pre_patch_object_claroline_corebundle_entity_user, method: prePatch }
             - { name: kernel.event_listener, event: crud_post_patch_object_claroline_corebundle_entity_user, method: postPatch }
+            - { name: kernel.event_listener, event: crud_pre_delete_object_claroline_corebundle_entity_user, method: preDelete }
+            - { name: kernel.event_listener, event: crud_post_delete_object_claroline_corebundle_entity_user, method: postDelete }
 
     Claroline\CoreBundle\API\Crud\User\RoleCrud:
         arguments:
@@ -84,6 +87,7 @@ services:
             - '@security.token_storage'
             - '@Claroline\AppBundle\Persistence\ObjectManager'
             - '@Claroline\AppBundle\API\Crud'
+            - '@Claroline\CoreBundle\Manager\FileManager'
             - '@Claroline\CoreBundle\Manager\Resource\ResourceLifecycleManager'
             - '@Claroline\CoreBundle\Manager\ResourceManager'
             - '@Claroline\CoreBundle\Manager\Resource\RightsManager'
