+++ src/main/core/Resources/config/services/serializer.yml
@@ -43,7 +43,6 @@ services:
             - '@security.token_storage'
             - '@Claroline\AppBundle\Persistence\ObjectManager'
             - '@claroline.manager.workspace_manager'
-            - '@Claroline\CoreBundle\Library\Utilities\FileUtilities'
             - '@Claroline\CoreBundle\API\Serializer\User\UserSerializer'
             - '@Claroline\CoreBundle\API\Serializer\User\OrganizationSerializer'
             - '@Claroline\CoreBundle\API\Serializer\File\PublicFileSerializer'
