--- src/plugin/announcement/Resources/config/services/crud.yml
@@ -1,21 +1,0 @@
-services:
-    _defaults:
-        autowire: false
-        autoconfigure: false
-        public: false
-
-    Claroline\AnnouncementBundle\Crud\AnnouncementAggregateCrud:
-        arguments:
-            - '@Claroline\AppBundle\API\Crud'
-        tags:
-            - { name: kernel.event_listener, event: crud_pre_delete_object_claroline_announcementbundle_entity_announcementaggregate, method: preDelete }
-
-    Claroline\AnnouncementBundle\Crud\AnnouncementCrud:
-        arguments:
-            - '@security.token_storage'
-            - '@Claroline\AppBundle\Persistence\ObjectManager'
-            - '@Claroline\AnnouncementBundle\Manager\AnnouncementManager'
-        tags:
-            - { name: kernel.event_listener, event: crud_pre_create_object_claroline_announcementbundle_entity_announcement, method: preCreate }
-            - { name: kernel.event_listener, event: crud_post_update_object_claroline_announcementbundle_entity_announcement, method: postUpdate }
-            - { name: kernel.event_listener, event: crud_pre_delete_object_claroline_announcementbundle_entity_announcement, method: preDelete }
