+++ src/plugin/announcement/Resources/config/services/listener.yml
@@ -23,12 +23,3 @@ services:
             - { name: kernel.event_listener, event: resource.claroline_announcement_aggregate.copy, method: copy }
             - { name: kernel.event_listener, event: transfer.claroline_announcement_aggregate.export, method: onExport }
             - { name: kernel.event_listener, event: transfer.claroline_announcement_aggregate.import.after, method: onImport }
-
-    Claroline\AnnouncementBundle\Listener\CrudListener:
-        arguments:
-            - '@Claroline\AppBundle\Event\StrictDispatcher'
-        tags:
-            - { name: kernel.event_listener, event: crud.post.create.announcement, method: onAnnouncementCreate, priority: -1 }
-            - { name: kernel.event_listener, event: crud.post.create.announcement_send, method: onAnnouncementSend }
-            - { name: kernel.event_listener, event: crud.post.update.announcement, method: onAnnouncementUpdate }
-            - { name: kernel.event_listener, event: crud.pre.delete.announcement, method: onAnnouncementDelete }
