+++ src/plugin/announcement/Resources/config/services/serializer.yml
@@ -13,7 +13,6 @@ services:
             - '@Claroline\CoreBundle\API\Serializer\Workspace\WorkspaceSerializer'
             - '@Claroline\CoreBundle\API\Serializer\Resource\ResourceNodeSerializer'
             - '@Claroline\CoreBundle\API\Serializer\File\PublicFileSerializer'
-            - '@Claroline\CoreBundle\Library\Utilities\FileUtilities'
             - '@Claroline\CoreBundle\API\Serializer\User\RoleSerializer'
 
     Claroline\AnnouncementBundle\Serializer\AnnouncementAggregateSerializer:
