+++ src/plugin/announcement/Subscriber/Crud/AnnouncementSubscriber.php
@@ -1,40 +1,65 @@
 <?php
 
-namespace Claroline\AnnouncementBundle\Crud;
+namespace Claroline\AnnouncementBundle\Subscriber\Crud;
 
 use Claroline\AnnouncementBundle\Entity\Announcement;
 use Claroline\AnnouncementBundle\Entity\AnnouncementSend;
+use Claroline\AnnouncementBundle\Event\Log\LogAnnouncementEvent;
 use Claroline\AnnouncementBundle\Manager\AnnouncementManager;
+use Claroline\AppBundle\API\Crud;
 use Claroline\AppBundle\API\Options;
 use Claroline\AppBundle\Event\Crud\CreateEvent;
 use Claroline\AppBundle\Event\Crud\DeleteEvent;
 use Claroline\AppBundle\Event\Crud\UpdateEvent;
+use Claroline\AppBundle\Event\StrictDispatcher;
 use Claroline\AppBundle\Persistence\ObjectManager;
 use Claroline\CoreBundle\Entity\User;
 use Claroline\CoreBundle\Library\Normalizer\DateNormalizer;
+use Claroline\CoreBundle\Manager\FileManager;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
-class AnnouncementCrud
+class AnnouncementSubscriber implements EventSubscriberInterface
 {
     /** @var TokenStorageInterface */
     private $tokenStorage;
+    /** @var StrictDispatcher */
+    private $dispatcher;
     /** @var ObjectManager */
     private $om;
     /** @var AnnouncementManager */
     private $manager;
+    /** @var FileManager */
+    private $fileManager;
 
     public function __construct(
         TokenStorageInterface $tokenStorage,
+        StrictDispatcher $dispatcher,
         ObjectManager $om,
-        AnnouncementManager $manager
+        AnnouncementManager $manager,
+        FileManager $fileManager
     ) {
         $this->tokenStorage = $tokenStorage;
+        $this->dispatcher = $dispatcher;
         $this->om = $om;
         $this->manager = $manager;
+        $this->fileManager = $fileManager;
+    }
+
+    public static function getSubscribedEvents()
+    {
+        return [
+            Crud::getEventName('create', 'pre', Announcement::class) => 'preCreate',
+            Crud::getEventName('create', 'post', Announcement::class) => 'postCreate',
+            Crud::getEventName('update', 'post', Announcement::class) => 'postUpdate',
+            Crud::getEventName('delete', 'pre', Announcement::class) => 'preDelete',
+            Crud::getEventName('delete', 'post', Announcement::class) => 'postDelete',
+        ];
     }
 
     public function preCreate(CreateEvent $event)
     {
+        /** @var Announcement $announcement */
         $announcement = $event->getObject();
         $options = $event->getOptions();
         $announcement->setAggregate($options['announcement_aggregate']);
@@ -48,11 +73,24 @@ class AnnouncementCrud
         }
     }
 
+    public function postCreate(CreateEvent $event)
+    {
+        /** @var Announcement $announcement */
+        $announcement = $event->getObject();
+
+        if ($announcement->getPoster()) {
+            $this->fileManager->linkFile(Announcement::class, $announcement->getUuid(), $announcement->getPoster());
+        }
+
+        $this->dispatchAnnouncementEvent($announcement, 'announcement-create');
+    }
+
     public function postUpdate(UpdateEvent $event)
     {
         /** @var Announcement $announcement */
         $announcement = $event->getObject();
         $data = $event->getData();
+        $oldData = $event->getOldData();
 
         // manage announce sending
         if (!empty($data['meta']) && !empty($data['meta']['notifyUsers']) && !empty($announcement->getRoles())) {
@@ -66,10 +104,20 @@ class AnnouncementCrud
                     break;
             }
         }
+
+        $this->fileManager->updateFile(
+            Announcement::class,
+            $announcement->getUuid(),
+            $announcement->getPoster(),
+            !empty($oldData['poster']) ? $oldData['poster']['url'] : null
+        );
+
+        $this->dispatchAnnouncementEvent($announcement, 'announcement-update');
     }
 
     public function preDelete(DeleteEvent $event)
     {
+        /** @var Announcement $announcement */
         $announcement = $event->getObject();
 
         if (!in_array(Options::SOFT_DELETE, $event->getOptions())) {
@@ -79,7 +127,27 @@ class AnnouncementCrud
             }
         }
 
-        // delete scheduled task is any
+        // delete scheduled task if any
         $this->manager->unscheduleMessage($announcement);
+
+        $this->dispatchAnnouncementEvent($announcement, 'announcement-delete');
+    }
+
+    public function postDelete(DeleteEvent $event)
+    {
+        /** @var Announcement $announcement */
+        $announcement = $event->getObject();
+
+        if (!in_array(Options::SOFT_DELETE, $event->getOptions()) && $announcement->getPoster()) {
+            $this->fileManager->unlinkFile(Announcement::class, $announcement->getUuid(), $announcement->getPoster());
+        }
+    }
+
+    /**
+     * @deprecated
+     */
+    private function dispatchAnnouncementEvent(Announcement $announcement, $action)
+    {
+        $this->dispatcher->dispatch('log', LogAnnouncementEvent::class, [$announcement, $action]);
     }
 }
