--- src/plugin/forum/Resources/config/services/crud.yml
@@ -1,25 +1,0 @@
-services:
-    _defaults:
-        autowire: false
-        autoconfigure: false
-        public: true
-
-    Claroline\ForumBundle\Crud\SubjectCrud:
-        arguments:
-            - '@Claroline\AppBundle\Persistence\ObjectManager'
-            - '@Claroline\ForumBundle\Manager\ForumManager'
-            - '@security.authorization_checker'
-            - '@messenger.default_bus'
-        tags:
-            - { name: kernel.event_listener, event: crud_pre_create_object_claroline_forumbundle_entity_subject, method: preCreate }
-            - { name: kernel.event_listener, event: crud_post_create_object_claroline_forumbundle_entity_subject, method: postCreate }
-
-    Claroline\ForumBundle\Crud\MessageCrud:
-        arguments:
-            - '@Claroline\AppBundle\Persistence\ObjectManager'
-            - '@Claroline\ForumBundle\Manager\ForumManager'
-            - '@security.authorization_checker'
-            - '@messenger.default_bus'
-        tags:
-            - { name: kernel.event_listener, event: crud_pre_create_object_claroline_forumbundle_entity_message, method: preCreate }
-            - { name: kernel.event_listener, event: crud_post_create_object_claroline_forumbundle_entity_message, method: postCreate }
