--- src/plugin/open-badge/Resources/config/services/crud.yml
@@ -1,13 +1,0 @@
-services:
-    _defaults:
-        autowire: false
-        autoconfigure: false
-        public: true #because lazy
-
-    Claroline\OpenBadgeBundle\Crud\BadgeClassCrud:
-        arguments:
-            - '@security.token_storage'
-            - '@Claroline\CoreBundle\Manager\Organization\OrganizationManager'
-        tags:
-            - { name: kernel.event_listener, event: crud_pre_create_object_claroline_openbadgebundle_entity_badgeclass, method: checkOrganization }
-            - { name: kernel.event_listener, event: crud_pre_update_object_claroline_openbadgebundle_entity_badgeclass, method: checkOrganization }
