+++ src/plugin/open-badge/Resources/config/services/serializer.yml
@@ -18,7 +18,6 @@ services:
         tags: [claroline.serializer]
         arguments:
             - '@security.authorization_checker'
-            - '@Claroline\CoreBundle\Library\Utilities\FileUtilities'
             - '@router'
             - '@Claroline\AppBundle\Persistence\ObjectManager'
             - '@Claroline\OpenBadgeBundle\Serializer\CriteriaSerializer'
