+++ src/plugin/open-badge/Resources/config/services/subscriber.yml
@@ -1,4 +1,18 @@
 services:
+    Claroline\OpenBadgeBundle\Subscriber\Crud\BadgeClassSubscriber:
+        arguments:
+            - '@security.token_storage'
+            - '@Claroline\CoreBundle\Manager\Organization\OrganizationManager'
+            - '@Claroline\CoreBundle\Manager\FileManager'
+        tags:
+            - { name: kernel.event_subscriber }
+
+    Claroline\OpenBadgeBundle\Subscriber\Crud\WorkspaceSubscriber:
+        arguments:
+            - '@Claroline\AppBundle\Persistence\ObjectManager'
+            - '@Claroline\AppBundle\API\Crud'
+        tags:
+            - { name: kernel.event_subscriber }
 
     Claroline\OpenBadgeBundle\Subscriber\BadgeLogSubscriber:
         arguments:
