+++ src/plugin/slideshow/Resources/config/services/listener.yml
@@ -7,5 +7,7 @@ services:
     Claroline\SlideshowBundle\Listener\Resource\SlideshowListener:
         arguments:
             - '@Claroline\AppBundle\API\SerializerProvider'
+            - '@Claroline\CoreBundle\Manager\FileManager'
         tags:
             - { name: kernel.event_listener, event: resource.claro_slideshow.load, method: onLoad }
+            - { name: kernel.event_listener, event: resource.claro_slideshow.delete, method: onDelete }
